[SQONE-6369] Added endpoints to list OOB assets
